% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build.r
\name{build_topics}
\alias{build_topics}
\title{Generate all topic pages for a package.}
\usage{
build_topics(pkg = ".", topics, site_path = NULL, templates_path = NULL)
}
\description{
Generate all topic pages for a package.
}

