% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to-rst.r
\name{to_rst}
\alias{to_rst}
\title{Convert an rdoc to a list of rst components.}
\usage{
to_rst(x, ...)
}
\arguments{
\item{x}{rd object to convert to rst}

\item{...}{other arguments passed onto to methods}
}
\value{
A list, suitable for rendering with
  \code{\link[whisker]{whisker.render}}
}
\description{
All span-level tags are converted to rst, and higher level blocks are
returned as components of the list.
}

